requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap-v5==1.0.5
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-responsive==0.3.0
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==3.10.1
Pillow==8.3.2
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.5.0
